deleted about page
