+++ src/components/UserPage/UserPage.jsx
@@ -1,16 +1,11 @@
-import React, { useEffect } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
+import React from 'react';
+import { useSelector} from 'react-redux';
 
 //  This page is for users to view all events they subscribed to and edit their profile info.
 function UserPage() {
 
-  const dispatch = useDispatch();
-  const store = useSelector(store => store);
   const user = useSelector((store) => store.user);
 
-  useEffect(() => {
-    dispatch({ type: 'FETCH_EVENT' });
-  }, [])
 
   return (
     <>
@@ -19,12 +14,7 @@ function UserPage() {
         <h2>Welcome, {user.username}!</h2>
         <p>Your ID is: {user.id}</p>
       </div>
-      <div>
-        <ul>
-          {store.event[0] && store.event.map((event) => 
-          <li key={event.id}>{event.name}, {event.description}</li>)}
-        </ul>
-      </div>
+
     </>
   );
 }
